api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
kafka-python==2.0.2
pydantic==2.8.2
prometheus-client==0.20.0
psycopg2-binary==2.9.9
requests==2.32.3
